Moved base flux initialization to State constructor
