+++ benchmarks/cyl_rot/README.md
@@ -1,21 +1,6 @@
-# Benchmarks
+# NEEDS TO BE UPDATED!!!
 
-These should be relatively static tests to make sure we get consistent results as we change the code with the same grid and time-stepping for all implementations
-
-### Re=40 cylinder (steady)
-Compared against C++ and MATLAB implementations on benchmark case from Colonius & Taira (2008) with $n_b=78$ points on the cylinder with steady free-stream flow.
-Script to run the validation case is `cyl40.jl`.
-
-| Package      | Runtime (secs) |  $C_D$ |
-| ----------- | ----------- | -----|
-| MATLAB      | 700       |  1.5738  |
-| C++   | 1164      | 1.5320  |
-| Fortran |  865  |  1.5738  |
-| Julia (single-core)  | 691   |  1.5738 |
-| Julia (4 threads)  |  488   |  1.5738  |
-
-
-### Re=20 cylinder (rotating)
+# Re=20 cylinder (rotating)
 Compare different magainst C++ implementation (lab and body-fixed frames).
 Rotation is $\Omega = 0.1$ rad/s, which gives a theoretical prediction of $C_L = -0.157$ according to Kutta-Joukouski
 There are three different approaches to treating the cylinder:
