+++ src/pre-processing/pre-processing-include.jl
@@ -3,7 +3,7 @@ Information about which files to include within this directory
 """
 
 #include("grid.jl")    # Moved to fluid-domain
-include("motions.jl")  # Move to structure-domain?
-include("bodies.jl")   # Move to structure-domain?
+#include("motions.jl")  # Moved to structure-domain
+#include("bodies.jl")   # Moved to structure-domain
 include("models.jl")          # model, matrices, working memory
-include("problem-types.jl")   # problem and state
+include("problem-types.jl")   # problem, state, and base flux
