+++ src/pre-processing/problem-types.jl
@@ -76,6 +76,123 @@ mutable struct IBState{T<:Grid} <: State
         state.CD = zeros(length(prob.model.bodies))
         state.CL = zeros(length(prob.model.bodies))
         state.cfl, state.slip = 0.0, 0.0
+        base_flux!(state, prob, 0.0)  # Initialize base flux at time zero
         return state
+    end
+end
+
+
+"""
+    base_flux!(state::IBState, prob::IBProblem, t::Float64)
+Set background flux based on `prob.model.bodies[].motion`
+Assumes same free-stream parameters for all motions (<-- CHANGE THIS)
+"""
+function base_flux!(state::IBState,
+                    prob::IBProblem,
+                    t::Float64)
+    base_flux!(MotionType(prob.model.bodies), state, prob, t)
+end
+
+"Initialize irrotational freestream flux when not time-varying"
+function base_flux!(::Type{T} where T <: InertialMotion,
+                    state::IBState{UniformGrid},
+                    prob::IBProblem,
+                    t::Float64)
+    grid = prob.model.grid
+    Uinf, α = prob.model.Uinf, prob.model.α
+    m = grid.nx;
+    n = grid.ny;
+    state.q0[ 1:(m-1)*n ] .= Uinf * grid.h * cos(α);  # x-flux
+    state.q0[ (m-1)*n+1:end ] .= Uinf * grid.h * sin(α);  # y-flux
+end
+
+"Initialize irrotational freestream flux when not time-varying"
+function base_flux!(::Type{T} where T <: InertialMotion,
+                    state::IBState{MultiGrid},
+                    prob::IBProblem,
+                    t::Float64)
+    grid = prob.model.grid
+    Uinf, α = prob.model.Uinf, prob.model.α
+    m = grid.nx;
+    n = grid.ny;
+    for lev = 1 : grid.mg
+        # Coarse grid spacing
+        hc = grid.h * 2^( lev - 1 );
+
+        # write fluid velocity flux in body-fixed frame
+        state.q0[ 1:(m-1)*n, lev ] .= Uinf * hc * cos(α);      # x-flux
+        state.q0[ (m-1)*n+1:end, lev ] .= Uinf * hc * sin(α);  # y-flux
+    end
+end
+
+"Update time-varying background flux for moving grid"
+function base_flux!(::Type{MovingGrid},
+                    state::IBState{UniformGrid},
+                    prob::IBProblem,
+                    t::Float64)
+    @assert length(prob.model.bodies) == 1 # Assumes only one body
+    grid = prob.model.grid
+    motion = prob.model.bodies[1].motion
+    nu = grid.ny*(grid.nx-1);  # Number of x-flux points
+
+    ### Rotational part
+    Ω = -motion.θ̇(t)
+    α = -motion.θ(t)
+    nx, ny, h = grid.nx, grid.ny, grid.h;
+
+    ### x-flux
+    # TODO: CHECK OFFSETS FOR THESE
+    y = h*(1:ny) .- grid.offy
+    YY = ones(nx-1)*y'
+    state.q0[1:nu, 1] .= -h*Ω*YY[:]
+
+    ### y-flux
+    x = h*(1:nx) .- grid.offx
+    XX = x*ones(ny-1)'
+    state.q0[nu+1:grid.nq, 1] .= h*Ω*XX[:]
+
+    ### Potential flow part
+    Ux0 = motion.U(t)*cos(α)
+    Uy0 = motion.U(t)*sin(α)
+    state.q0[1:nu, 1] .+= h*Ux0          # x-flux
+    state.q0[nu+1:grid.nq, 1] .+= h*Uy0  # y-velocity
+end
+
+"Update time-varying background flux for moving grid"
+function base_flux!(::Type{MovingGrid},
+                    state::IBState{MultiGrid},
+                    prob::IBProblem,
+                    t::Float64)
+    @assert length(prob.model.bodies) == 1 # Assumes only one body
+    grid = prob.model.grid
+    motion = prob.model.bodies[1].motion
+    XX, YY = prob.model.XX, prob.model.YY;
+    nu = grid.ny*(grid.nx-1);  # Number of x-flux points
+
+    ### Rotational part
+    Ω = -motion.θ̇(t)
+    α = -motion.θ(t)
+
+    ### Potential flow part (note θ = -α for angle of attack)
+    Ux0 = motion.U(t)*cos(α)
+    Uy0 = motion.U(t)*sin(α)
+
+    state.q0 .*= 0.0
+    for lev=1:grid.mg
+        hc = grid.h*2^(lev-1);  # Coarse grid spacing
+
+        ### x-fluxes
+        #y = @. ((1:ny)-0.5-ny/2)*hc + ny/2*h - grid.offy
+        #YY = ones(nx-1)*y'   # TODO:  Pre-compute or find a better way to do this
+        state.q0[1:nu, lev] .= -hc*Ω*YY[:, lev]
+
+        ### y-fluxes
+        #x = @. ((1:nx)-0.5-nx/2)*hc + nx/2*h - grid.offx
+        #XX = x*ones(ny-1)'   # TODO:  Pre-compute or find a better way to do this
+        state.q0[nu+1:end, lev] .= hc*Ω*XX[:, lev]
+
+        ### Irrotational part
+        state.q0[1:nu, lev] .+= hc*Ux0          # x-flux
+        state.q0[nu+1:end, lev] .+= hc*Uy0  # y-velocity
     end
 end
